First Commit.
For now, I made a small change at total cost calculation (Separated the values for easier debugging)

The problem with the obstacle avoidance seems the lack of normalization in the cost calculation. I might need to add a function for this, or add it straight in the cost calculation function itself.

Is this text too big to be in a commit?